--- satellite/marketing/marketingweb/static/pages/404.html
@@ -1,3 +1,0 @@
-{{define "main"}}
-    <h1 class="m-5 text-center">404<br>Not Found</h1>
-{{end}}
