--- satellite/marketing/marketingweb/static/pages/banner.html
@@ -1,14 +1,0 @@
-{{define "banner"}}
-<div class="container-fluid banner bg-white w-100">
-    <div class="container">
-        <div class="row d-flex justify-content-between">
-            <div class="mt-4">
-                <a class="home-link" href="/">
-                    <span class="ml-3">{{template "logo" .}}</span>
-                </a>
-            </div>
-            <p class="ref-admin mr-3">Referral Admin</h5>
-        </div>
-    </div>
-</div>
-{{end}}
