--- satellite/marketing/marketingweb/static/pages/freeOffers.html
@@ -1,46 +1,0 @@
-{{define "freeOffers"}}
-<div class="offers-table mt-2 mb-5 container">
-        <div class="row data-row">
-            <div class="col col-heading">Name</div>
-            <div class="col col-heading">Give Credit</div>
-            <div class="col col-heading">Referrals&nbsp;Used</div> 
-            <div class="col col-heading">Redeemable&nbsp;Capacity</div>
-            <div class="col col-heading">Created</div>
-            <div class="col col-heading">Expiration</div>
-            <div class="col col-heading">Status</div>
-            <div class="col col-heading edit-offer"></div>
-        </div><hr>
-        <div class="row offer-heading ">
-            <p class="offer-type">DEFAULT&nbsp;OFFER</p>
-        </div>
-        {{range .FreeCredits}}
-            {{if eq .Status 2}}
-                <div class="row data-row">
-                    <div class="col ml-3">{{.Name}}</div>
-                    <div class="col">${{.InviteeCreditInCents}}</div>
-                    <div class="col">{{.NumRedeemed}}</div> 
-                    <div class="col">{{.RedeemableCap}}</div>
-                    <div class="col">{{printf "%.10s" .CreatedAt}}</div>
-                    <div class="col">{{printf "%.10s" .ExpiresAt}}</div>
-                    <div class="col"><a href="#">{{if eq .Status 1}}<strong>Live &#183;</strong>{{else}}off{{end}}</a></div>
-                    <div class="col col-heading edit-offer">EDIT</div>
-                </div><hr>
-            {{end}}
-        {{end}}
-        <div class="row offer-heading ">
-            <p class="offer-type">CUSTOM&nbsp;OFFERS</p>
-        </div>
-        {{range .FreeCredits}}
-        <div class="row data-row">
-            <div class="col ml-3">{{.Name}}</div>
-            <div class="col">${{.InviteeCreditInCents}}</div> 
-            <div class="col">0</div>
-            <div class="col">{{.RedeemableCap}}</div>
-            <div class="col">{{printf "%.10s" .CreatedAt}}</div>
-            <div class="col">{{printf "%.10s" .ExpiresAt}}</div>
-            <div class="col"><a href="#">{{if eq .Status 1}}<strong>Live &#183;</strong>{{else}}off{{end}}</a></div>
-            <div class="col col-heading edit-offer">EDIT</div>
-        </div><hr>
-        {{end}}
-    </div>
-{{end}}
