+++ web/marketing/pages/home.html
@@ -7,13 +7,13 @@ See LICENSE for copying information. -->
         <h3 class="mt-4">Referral Credit Offers</h3>
         <button type="button" class="btn btn-primary mt-4" data-toggle="modal" data-target="#ref-offers-modal">+ Create Referral Credit</button>
     </div>
-    {{template "refOffers" .}}
-    {{template "refOffersModal" .}}
+    {{template "referral-offers" .}}
+    {{template "referral-Offers-modal" .}}
     <div class="row mt-4 mb-4 d-flex justify-content-between">
         <h3 class="mt-4">Free Credit Offers</h3>
         <button type="button" class="btn btn-primary mt-4" data-toggle="modal" data-target="#free-offers-modal">+ Create Free Credit</button>
     </div>
-    {{template "freeOffers" .}}
-    {{template "freeOffersModal" .}}
+    {{template "free-offers" .}}
+    {{template "free-offers-modal" .}}
 </div>
 {{end}}
