+++ web/marketing/pages/internal-server-error.html
@@ -2,5 +2,6 @@
 See LICENSE for copying information. -->
 
 {{define "main"}}
-    <h1>Internal Server Error</h1>
+    <h1 class="m-5 text-center">Internal Server Error</h1>
+    <h2 class="m-5 text-center">{{.}}</h2>
 {{end}}
